Backend/requirements.txt
nfcpy>=1.0.4
requests>=2.31.0
pytest>=7.4.0
